training/pyproject.toml
[project]
name = "emu-training"
version = "0.1.0"
description = "Synthetic trajectory generation for emu harness SFT (OSWorld → emu-style)"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "huggingface-hub>=0.25.0",
    "remotezip>=0.12.2",
]

[tool.uv]
package = false
